+++ srcs/ft_printf_args.c
@@ -6,17 +6,18 @@
 /*   By: mbrignol <mbrignol@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/21 15:32:12 by mbrignol          #+#    #+#             */
-/*   Updated: 2019/12/01 02:25:47 by mbrignol         ###   ########.fr       */
+/*   Updated: 2019/12/02 18:43:38 by mbrignol         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/ft_printf.h"
+#include "../includes/ft_printf.h"
 
-int	get_arg_type(va_list va, t_flag *info)
+int		get_arg_type(va_list va, t_flag *info)
 {
 	if (info->letter == 'c')
 		return (get_char_arg(va, info));
-	if (info->letter == 'd' || info->letter == 'i' || info->letter == 'x' || info->letter == 'X' || info->letter == 'u' || info->letter == 'p')
+	if (info->letter == 'd' || info->letter == 'i' || info->letter == 'x'
+		|| info->letter == 'X' || info->letter == 'u' || info->letter == 'p')
 		return (get_int_arg(va, info));
 	if (info->letter == 's')
 		return (get_string_arg(va, info));
@@ -43,9 +44,8 @@ int		get_int_arg(va_list va, t_flag *info)
 	else if (info->letter == 'u')
 		nbr = ft_itoa_unsigned(va_arg(va, unsigned int));
 	else if (info->letter == 'X' || info->letter == 'x' || info->letter == 'p')
-		 nbr = ft_itoa_base(va_arg(va, int), 16, info);
+		nbr = ft_itoa_base(va_arg(va, long int), 16, info);
 	info->str = nbr;
-	//free(nbr);
 	return (1);
 }
 
@@ -55,9 +55,7 @@ int		get_char_arg(va_list va, t_flag *info)
 
 	s = ft_strnew(1);
 	s[0] = (char)va_arg(va, int);
-
 	info->str = s;
-	//free(s);
 	return (1);
 }
 
@@ -67,8 +65,6 @@ int		get_percent_arg(t_flag *info)
 
 	s = ft_strnew(1);
 	s[0] = '%';
-
 	info->str = s;
-	//free(s);
 	return (1);
 }
